Move balance table into market, and some constants closer to their use
