+++ actors/market/Cargo.toml
@@ -15,6 +15,7 @@ crate-type = ["cdylib", "lib"]
 
 [dependencies]
 fil_actors_runtime = { version = "7.0.0", path = "../runtime" }
+fvm_ipld_hamt = "0.2.0"
 fvm_shared = { version = "0.2.0", default-features = false }
 bitfield = { version = "0.2.0", package = "fvm_ipld_bitfield" }
 num-traits = "0.2.14"
