+++ actors/market/tests/market_actor_test.rs
@@ -3,16 +3,15 @@
 
 use std::collections::HashMap;
 
+use fil_actor_market::balance_table::{BalanceTable, BALANCE_TABLE_BITWIDTH};
 use fil_actor_market::{
     ext, Actor as MarketActor, Method, State, WithdrawBalanceParams, PROPOSALS_AMT_BITWIDTH,
     STATES_AMT_BITWIDTH,
 };
 use fil_actors_runtime::runtime::Runtime;
 use fil_actors_runtime::test_utils::*;
-use fil_actors_runtime::util::BALANCE_TABLE_BITWIDTH;
 use fil_actors_runtime::{
-    make_empty_map, ActorError, BalanceTable, SetMultimap, STORAGE_MARKET_ACTOR_ADDR,
-    SYSTEM_ACTOR_ADDR,
+    make_empty_map, ActorError, SetMultimap, STORAGE_MARKET_ACTOR_ADDR, SYSTEM_ACTOR_ADDR,
 };
 use fvm_ipld_amt::Amt;
 use fvm_shared::address::Address;
