+++ actors/power/src/state.rs
@@ -6,8 +6,8 @@ use std::ops::Neg;
 use anyhow::{anyhow, Context};
 use cid::Cid;
 use fil_actors_runtime::{
-    actor_error, consensus_miner_min_power, make_empty_map, make_map_with_root,
-    make_map_with_root_and_bitwidth, ActorDowncast, ActorError, Map, Multimap,
+    actor_error, make_empty_map, make_map_with_root, make_map_with_root_and_bitwidth,
+    ActorDowncast, ActorError, Map, Multimap,
 };
 use fvm_ipld_hamt::BytesKey;
 use fvm_shared::address::Address;
@@ -398,6 +398,33 @@ pub struct CronEvent {
 }
 
 impl Cbor for CronEvent {}
+
+/// Returns the minimum storage power required for each seal proof types.
+pub fn consensus_miner_min_power(p: RegisteredPoStProof) -> anyhow::Result<StoragePower> {
+    use RegisteredPoStProof::*;
+    match p {
+        StackedDRGWinning2KiBV1
+        | StackedDRGWinning8MiBV1
+        | StackedDRGWinning512MiBV1
+        | StackedDRGWinning32GiBV1
+        | StackedDRGWinning64GiBV1
+        | StackedDRGWindow2KiBV1
+        | StackedDRGWindow8MiBV1
+        | StackedDRGWindow512MiBV1
+        | StackedDRGWindow32GiBV1
+        | StackedDRGWindow64GiBV1 => {
+            let power: u64 = if cfg!(feature = "min-power-2k") {
+                2 << 10
+            } else if cfg!(feature = "min-power-2g") {
+                2 << 30
+            } else {
+                10 << 40
+            };
+            Ok(StoragePower::from(power))
+        }
+        Invalid(i) => Err(anyhow::anyhow!("unsupported proof type: {}", i)),
+    }
+}
 
 #[cfg(test)]
 mod test {
