+++ actors/runtime/src/builtin/mod.rs
@@ -2,11 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 pub use self::network::*;
-pub use self::sector::*;
 pub use self::shared::*;
 pub use self::singletons::*;
 
 pub mod network;
-pub mod sector;
 pub mod shared;
 pub mod singletons;
