+++ actors/runtime/src/builtin/network.rs
@@ -1,7 +1,6 @@
 // Copyright 2019-2022 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use fvm_shared::bigint::BigInt;
 use fvm_shared::clock::EPOCH_DURATION_SECONDS;
 pub use fvm_shared::BLOCKS_PER_EPOCH as EXPECTED_LEADERS_PER_EPOCH;
 
@@ -11,17 +10,3 @@ pub const SECONDS_IN_YEAR: i64 = 31556925;
 pub const EPOCHS_IN_HOUR: i64 = SECONDS_IN_HOUR / EPOCH_DURATION_SECONDS;
 pub const EPOCHS_IN_DAY: i64 = SECONDS_IN_DAY / EPOCH_DURATION_SECONDS;
 pub const EPOCHS_IN_YEAR: i64 = SECONDS_IN_YEAR / EPOCH_DURATION_SECONDS;
-
-/// Precision used for making QA power calculations
-pub const SECTOR_QUALITY_PRECISION: i64 = 20;
-
-lazy_static! {
-    /// Quality multiplier for committed capacity (no deals) in a sector
-    pub static ref QUALITY_BASE_MULTIPLIER: BigInt = BigInt::from(10);
-
-    /// Quality multiplier for unverified deals in a sector
-    pub static ref DEAL_WEIGHT_MULTIPLIER: BigInt = BigInt::from(10);
-
-    /// Quality multiplier for verified deals in a sector
-    pub static ref VERIFIED_DEAL_WEIGHT_MULTIPLIER: BigInt = BigInt::from(100);
-}
