+++ actors/runtime/src/util/mod.rs
@@ -1,13 +1,11 @@
 // Copyright 2019-2022 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-pub use self::balance_table::{BalanceTable, BALANCE_TABLE_BITWIDTH};
 pub use self::downcast::*;
 pub use self::multimap::*;
 pub use self::set::Set;
 pub use self::set_multimap::SetMultimap;
 
-mod balance_table;
 pub mod chaos;
 mod downcast;
 mod multimap;
